refactor: Add commonly used types to prelude
changelog: ignore
